Run post details link navigation test on emulator
