+++ app/src/testDemo/java/br/com/orcinus/orca/app/demo/ProfileDetailsTests.kt
@@ -17,28 +17,16 @@ package br.com.orcinus.orca.app.demo
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onFirst
-import androidx.compose.ui.test.onLast
-import androidx.compose.ui.test.performClick
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import assertk.assertThat
 import br.com.orcinus.orca.app.R
 import br.com.orcinus.orca.app.activity.OrcaActivity
-import br.com.orcinus.orca.composite.timeline.test.isTimeline
-import br.com.orcinus.orca.composite.timeline.test.post.figure.link.onLinkCards
 import br.com.orcinus.orca.composite.timeline.test.post.onPostPreviews
-import br.com.orcinus.orca.composite.timeline.test.post.performScrollToPostPreviewWithLinkCard
-import br.com.orcinus.orca.core.auth.actor.Actor
-import br.com.orcinus.orca.core.sample.instance.SampleInstance
 import br.com.orcinus.orca.feature.postdetails.PostDetailsFragment
-import br.com.orcinus.orca.platform.core.image.sample
-import br.com.orcinus.orca.platform.core.sample
-import br.com.orcinus.orca.platform.intents.test.intendBrowsingTo
 import br.com.orcinus.orca.platform.navigation.test.isAt
-import br.com.orcinus.orca.std.image.compose.ComposableImageLoader
 import kotlin.test.BeforeTest
-import kotlinx.coroutines.test.runTest
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -51,29 +39,6 @@ internal class ProfileDetailsTests {
   @BeforeTest
   fun setUp() {
     onView(withId(R.id.profile_details)).perform(click())
-  }
-
-  @Test
-  fun navigatesToPostLink() {
-    runTest {
-      val instance =
-        SampleInstance.Builder.create(ComposableImageLoader.Provider.sample)
-          .withDefaultProfiles()
-          .withDefaultPosts()
-          .build()
-      intendBrowsingTo(
-        instance.postProvider
-          .provideAllCurrentBy(Actor.Authenticated.sample.id)
-          .firstNotNullOf { it.content.highlight }
-          .uri
-      ) {
-        composeRule
-          .onAllNodes(isTimeline())
-          .onLast()
-          .performScrollToPostPreviewWithLinkCard(instance.feedProvider)
-        composeRule.onLinkCards().onFirst().performClick()
-      }
-    }
   }
 
   @Test
